# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "NovoCore API",
    "version" : "0.1.0",
    "description" : "Generated from the live handler mapping by OpenApiSpecIT. Do not edit by hand — regenerate with `mvn verify -Dnovocore.openapi.write=true`.\n\nEvery monetary amount, unit cost, quantity and rate is a STRING, never a JSON number. Sending a number where one of these is expected is refused."
  },
  "paths" : {
...
    "/api/credit-notes" : {
      "get" : {
        "operationId" : "SalesController_notes",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "VIEW",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "salesInvoiceId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListResponse_CreditNoteView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "SalesController_issue",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCreditNote"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditNoteView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/credit-notes/preview" : {
      "post" : {
        "operationId" : "SalesController_previewNote",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewCreditNote"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditNotePreview"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/credit-notes/{id}" : {
      "get" : {
        "operationId" : "SalesController_note",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "VIEW",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditNoteView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/credit-notes/{id}/reversal" : {
      "post" : {
        "operationId" : "SalesController_reverseNote",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReversalCommand"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreditNoteView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sales-invoices" : {
      "get" : {
        "operationId" : "SalesController_invoices",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "VIEW",
        "parameters" : [ {
          "name" : "customerId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SalesInvoiceSort"
          }
        }, {
          "name" : "direction",
          "in" : "query",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListResponse_SalesInvoiceView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "operationId" : "SalesController_record",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewSalesInvoice"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoiceView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales-invoices/preview" : {
      "post" : {
        "operationId" : "SalesController_preview",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewSalesInvoice"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoicePreview"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales-invoices/rounding-differences" : {
      "get" : {
        "operationId" : "SalesController_roundingDifferences",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "VIEW",
        "parameters" : [ {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RoundingReport"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales-invoices/{id}" : {
      "get" : {
        "operationId" : "SalesController_invoice",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "VIEW",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoiceView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sales-invoices/{id}/reversal" : {
      "post" : {
        "operationId" : "SalesController_reverse",
        "x-novocore-section" : "SALES",
        "x-novocore-level" : "FULL",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReversalCommand"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SalesInvoiceView"
                }
              }
            }
          },
          "4XX" : {
            "description" : "The request was refused. RFC 7807 body with a 'detail' explaining why, except for permission refusals, which say nothing on purpose.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer"
                    },
                    "title" : {
                      "type" : "string"
                    },
                    "detail" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreditNotePreview" : {
        "type" : "object",
        "properties" : {
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreditNotePreviewLine"
            }
          },
          "net" : {
            "$ref" : "#/components/schemas/Money"
          },
          "vat" : {
            "$ref" : "#/components/schemas/Money"
          },
          "gross" : {
            "$ref" : "#/components/schemas/Money"
          },
          "statedTotal" : {
            "$ref" : "#/components/schemas/Money"
          },
          "roundingDifference" : {
            "$ref" : "#/components/schemas/Money"
          },
          "roundingThreshold" : {
            "$ref" : "#/components/schemas/Money"
          },
          "roundingNeedsAcceptance" : {
            "type" : "boolean"
          },
          "payable" : {
            "$ref" : "#/components/schemas/Money"
          }
        }
      },
...
      "ListResponse_CreditNoteView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CreditNoteView"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
...
      "ListResponse_SalesInvoiceView" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SalesInvoiceView"
            }
          },
          "page" : {
            "$ref" : "#/components/schemas/PageInfo"
          }
        }
      },
...
      "NewCreditNote" : {
        "type" : "object",
        "properties" : {
          "salesInvoiceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "documentNumber" : {
            "type" : "string"
          },
          "creditNoteDate" : {
            "type" : "string",
            "format" : "date"
          },
          "description" : {
            "type" : "string"
          },
          "statedTotal" : {
            "$ref" : "#/components/schemas/Money"
          },
          "roundingAcceptedBy" : {
            "type" : "string"
          },
          "roundingNote" : {
            "type" : "string"
          },
          "lines" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/NewCreditNoteLine"
            }
          }
        }
      },
...
      "ReversalCommand" : {
        "type" : "object",
        "properties" : {
          "reversalDate" : {
            "type" : "string",
            "format" : "date"
          },
          "reason" : {
            "type" : "string"
          }
        }
      },
...
      "RoundingReport" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SalesInvoiceView"
            }
          },
          "total" : {
            "$ref" : "#/components/schemas/Money"
          }
        }
      },
...
      "SalesInvoiceSort" : {
        "type" : "string",
        "enum" : [ "INVOICE_DATE", "DOCUMENT_NUMBER", "RECORDED_AT" ]
      },
...
      "SortDirection" : {
        "type" : "string",
        "enum" : [ "ASC", "DESC" ]
      },
...
    }
  }
}